requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
pandas
openpyxl
python-dotenv
email-validator
PyJWT
